# Improve performance for DictionaryBlock.getSizeInBytes