--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractFixedWidthBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractFixedWidthBlock.java
@@ -18,6 +18,7 @@ import io.airlift.slice.Slices;
 import io.airlift.slice.XxHash64;
 
 import static com.facebook.presto.spi.block.BlockUtil.checkValidRegion;
+import static com.facebook.presto.spi.block.BlockUtil.countUsedPositions;
 
 public abstract class AbstractFixedWidthBlock
         implements Block
@@ -184,6 +185,12 @@ public abstract class AbstractFixedWidthBlock
         int positionCount = getPositionCount();
         checkValidRegion(positionCount, positionOffset, length);
         return (fixedSize + Byte.BYTES) * (long) length;
+    }
+
+    @Override
+    public long getPositionsSizeInBytes(boolean[] positions)
+    {
+        return (fixedSize + Byte.BYTES) * (long) countUsedPositions(positions);
     }
 
     private int valueOffset(int position)