--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/BlockUtil.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/BlockUtil.java
@@ -49,6 +49,13 @@ final class BlockUtil
         }
     }
 
+    static void checkValidPositions(boolean[] positions, int positionCount)
+    {
+        if (positions.length != positionCount) {
+            throw new IllegalArgumentException(format("Invalid positions array size %d, actual position count is %d", positions.length, positionCount));
+        }
+    }
+
     static void checkValidPosition(int position, int positionCount)
     {
         if (position < 0 || position >= positionCount) {
@@ -172,6 +179,16 @@ final class BlockUtil
             return array;
         }
         return Arrays.copyOfRange(array, index, index + length);
+    }
+    static int countUsedPositions(boolean[] positions)
+    {
+        int used = 0;
+        for (boolean position : positions) {
+            if (position) {
+                used++;
+            }
+        }
+        return used;
     }
 
     /**