```
    Before:
    Benchmark                                              (selectedPositions)  Mode  Cnt       Score       Error  Units
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                  100  avgt   20    5173.612 ±  6164.744  us/op
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                 1000  avgt   20   32607.266 ± 37310.523  us/op
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                10000  avgt   20  113352.382 ± 13933.478  us/op
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes               100000  avgt   20  771976.351 ± 54696.701  us/op

    After:
    Benchmark                                              (selectedPositions)  Mode  Cnt     Score     Error  Units
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                  100  avgt   20   434.331 ±  49.481  us/op
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                 1000  avgt   20   612.484 ±  94.415  us/op
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                10000  avgt   20   908.409 ±  82.704  us/op
    BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes               100000  avgt   20  3043.432 ± 289.973  us/op
```