@rongrong 2018-11-08
Latest run with `@Fork(5)`:
# Run complete. Total time: 00:07:23

Benchmark                                              (selectedPositions)  Mode  Cnt     Score     Error  Units
BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                  100  avgt   50   348.270 ±  53.243  us/op
BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                 1000  avgt   50   677.579 ± 175.333  us/op
BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes                10000  avgt   50   874.065 ± 144.871  us/op
BenchmarkDictionaryBlockGetSizeInBytes.getSizeInBytes               100000  avgt   50  2610.320 ± 228.725  us/op